+++ src/tools/kata-ctl/src/ops/volume_ops.rs
@@ -8,12 +8,12 @@ use crate::args::{DirectVolSubcommand, DirectVolumeCommand};
 use anyhow::{anyhow, Ok, Result};
 use futures::executor;
 use kata_types::mount::{
-    DirectVolumeMountInfo, KATA_DIRECT_VOLUME_ROOT_PATH, KATA_MOUNT_INFO_FILE_NAME,
+    get_volume_mount_info, join_path, DirectVolumeMountInfo, KATA_DIRECT_VOLUME_ROOT_PATH,
+    KATA_MOUNT_INFO_FILE_NAME,
 };
 use nix;
 use reqwest::StatusCode;
-use safe_path;
-use std::{fs, path::PathBuf, time::Duration};
+use std::{fs, time::Duration};
 use url;
 
 use agent::ResizeVolumeRequest;
@@ -90,17 +90,6 @@ async fn stats(volume_path: &str) -> Result<Option<String>> {
     Ok(Some(body))
 }
 
-// join_path joins user provided volumepath with kata direct-volume root path
-// the volume_path is base64-encoded and then safely joined to the end of path prefix
-fn join_path(prefix: &str, volume_path: &str) -> Result<PathBuf> {
-    if volume_path.is_empty() {
-        return Err(anyhow!("volume path must not be empty"));
-    }
-    let b64_encoded_path = base64::encode(volume_path.as_bytes());
-
-    Ok(safe_path::scoped_join(prefix, b64_encoded_path)?)
-}
-
 // add writes the mount info (json string) of a direct volume into a filesystem path known to Kata Containers.
 pub fn add(volume_path: &str, mount_info: &str) -> Result<Option<String>> {
     let mount_info_dir_path = join_path(KATA_DIRECT_VOLUME_ROOT_PATH, volume_path)?;
@@ -127,15 +116,6 @@ pub fn remove(volume_path: &str) -> Result<Option<String>> {
     fs::remove_dir_all(path)?;
 
     Ok(None)
-}
-
-pub fn get_volume_mount_info(volume_path: &str) -> Result<DirectVolumeMountInfo> {
-    let mount_info_file_path =
-        join_path(KATA_DIRECT_VOLUME_ROOT_PATH, volume_path)?.join(KATA_MOUNT_INFO_FILE_NAME);
-    let mount_info_file = fs::read_to_string(mount_info_file_path)?;
-    let mount_info: DirectVolumeMountInfo = serde_json::from_str(&mount_info_file)?;
-
-    Ok(mount_info)
 }
 
 // get_sandbox_id_for_volume finds the id of the first sandbox found in the dir.
@@ -170,7 +150,7 @@ mod tests {
     use super::*;
     use kata_types::mount::DirectVolumeMountInfo;
     use serial_test::serial;
-    use std::{collections::HashMap, fs};
+    use std::{collections::HashMap, fs, path::PathBuf};
     use tempfile::tempdir;
     use test_utils::skip_if_not_root;
